+++ src/lib.rs
@@ -2,9 +2,8 @@
 #![cfg_attr(not(test), no_std)]
 
 /// Parser type with internal states.
-/// Owns a buffer of `SYSEX_MAX_LEN` bytes for constructing SysEx messages.
 #[derive(Debug)]
-pub struct MidiStreamParser<const SYSEX_MAX_LEN: usize> {
+pub struct MidiStreamParser {
     /// Buffer for message to be created.
     message: [u8; 3],
 
@@ -16,12 +15,16 @@ pub struct MidiStreamParser<const SYSEX_MAX_LEN: usize> {
 
     /// State of SysEx parsing.
     sysex_running: bool,
+}
 
-    /// SysEx message buffer.
-    sysex_message: [u8; SYSEX_MAX_LEN],
+/// Parser output variants.
+#[derive(Debug, PartialEq)]
+pub enum ParserOutput<'a> {
+    /// Regular message.
+    Message(&'a [u8]),
 
-    /// SysEx message length.
-    sysex_message_length: usize,
+    /// Byte of a SysEx message.
+    SysexByte(u8),
 }
 
 /// Error variants.
@@ -29,19 +32,16 @@ pub struct MidiStreamParser<const SYSEX_MAX_LEN: usize> {
 pub enum ParserError {
     /// No valid status byte.
     InvalidStatus,
-
-    /// SysEx message longer than SYSEX_MAX_LEN bytes.
-    SysexOverflow,
 }
 
-impl<const SYSEX_MAX_LEN: usize> Default for MidiStreamParser<SYSEX_MAX_LEN> {
+impl Default for MidiStreamParser {
     /// Returns a new parser with default values.
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl<const SYSEX_MAX_LEN: usize> MidiStreamParser<SYSEX_MAX_LEN> {
+impl MidiStreamParser {
     /// Returns a new parser.
     pub fn new() -> Self {
         Self {
@@ -49,24 +49,18 @@ impl<const SYSEX_MAX_LEN: usize> MidiStreamParser<SYSEX_MAX_LEN> {
             message_length: 0,
             realtime_message: [0; 1],
             sysex_running: false,
-            sysex_message: [0; SYSEX_MAX_LEN],
-            sysex_message_length: 0,
         }
     }
 
     /// Feed a byte into the parser and return result.
     /// The `Ok` variant is an option that contains either the constructed message or `None`
     /// in case the message is not ready yet.
-    pub fn parse(&mut self, byte: u8) -> Result<Option<&[u8]>, ParserError> {
+    pub fn parse<'a>(&'a mut self, byte: u8) -> Result<Option<ParserOutput<'a>>, ParserError> {
         match byte {
             0x00..=0x7F => {
                 // Data byte
                 if self.sysex_running {
-                    if self.sysex_message_length >= SYSEX_MAX_LEN {
-                        return Err(ParserError::SysexOverflow);
-                    }
-                    self.sysex_message[self.sysex_message_length] = byte;
-                    self.sysex_message_length += 1;
+                    return Ok(Some(ParserOutput::SysexByte(byte)));
                 } else {
                     if self.message_length == 0 {
                         // No valid status byte found.
@@ -77,13 +71,13 @@ impl<const SYSEX_MAX_LEN: usize> MidiStreamParser<SYSEX_MAX_LEN> {
                     if self.message_length == 3 {
                         // 3-byte message ready, keep first byte for running status
                         self.message_length = 1;
-                        return Ok(Some(&self.message));
+                        return Ok(Some(ParserOutput::Message(&self.message)));
                     } else if matches!(self.message[0] & 0xF0, 0xC0 | 0xD0)
                         || matches!(self.message[0], 0xF1 | 0xF3)
                     {
                         // 2-byte message ready, keep first byte for running status
                         self.message_length = 1;
-                        return Ok(Some(&self.message[0..2]));
+                        return Ok(Some(ParserOutput::Message(&self.message[0..2])));
                     }
                 }
             }
@@ -100,24 +94,18 @@ impl<const SYSEX_MAX_LEN: usize> MidiStreamParser<SYSEX_MAX_LEN> {
                         self.message[0] = 0;
                         self.message_length = 0;
                         self.sysex_running = true;
-                        self.sysex_message[0] = byte;
-                        self.sysex_message_length = 1;
+                        return Ok(Some(ParserOutput::SysexByte(byte)));
                     }
                     0xF6 => {
                         // Tune request.
                         self.message[0] = byte;
                         self.message_length = 1;
-                        return Ok(Some(&self.message[0..1]));
+                        return Ok(Some(ParserOutput::Message(&self.message[0..1])));
                     }
                     0xF7 => {
                         // End of SysEx.
                         self.sysex_running = false;
-                        if self.sysex_message_length >= SYSEX_MAX_LEN {
-                            return Err(ParserError::SysexOverflow);
-                        }
-                        self.sysex_message[self.sysex_message_length] = byte;
-                        self.sysex_message_length += 1;
-                        return Ok(Some(&self.sysex_message[0..self.sysex_message_length]));
+                        return Ok(Some(ParserOutput::SysexByte(byte)));
                     }
                     _ => {
                         self.message[0] = byte;
@@ -128,7 +116,7 @@ impl<const SYSEX_MAX_LEN: usize> MidiStreamParser<SYSEX_MAX_LEN> {
             0xF8..=0xFF => {
                 // Status byte for system realtime message.
                 self.realtime_message[0] = byte;
-                return Ok(Some(&self.realtime_message));
+                return Ok(Some(ParserOutput::Message(&self.realtime_message)));
             }
         }