+++ src/tests.rs
@@ -5,10 +5,10 @@ use super::*;
 /// Note off message.
 #[test]
 fn note_off() {
-    let mut parser = MidiStreamParser::<256>::new();
+    let mut parser = MidiStreamParser::new();
 
     let bytes = [0x82, 37, 10];
-    let messages = [None, None, Some([0x82, 37, 10].as_ref())];
+    let messages = [None, None, Some(ParserOutput::Message(&[0x82, 37, 10]))];
 
     for (byte, message) in bytes.iter().zip(messages.iter()) {
         let result = parser.parse(*byte).unwrap();
@@ -19,15 +19,15 @@ fn note_off() {
 /// Note off message with realtime messages in-between.
 #[test]
 fn note_off_with_realtime() {
-    let mut parser = MidiStreamParser::<256>::new();
+    let mut parser = MidiStreamParser::new();
 
     let bytes = [0x82, 0xFA, 37, 0xF8, 10];
     let messages = [
         None,
-        Some([0xFA].as_ref()),
+        Some(ParserOutput::Message(&[0xFA])),
         None,
-        Some([0xF8].as_ref()),
-        Some([0x82, 37, 10].as_ref()),
+        Some(ParserOutput::Message(&[0xF8])),
+        Some(ParserOutput::Message(&[0x82, 37, 10])),
     ];
 
     for (byte, message) in bytes.iter().zip(messages.iter()) {
@@ -39,10 +39,10 @@ fn note_off_with_realtime() {
 /// Note on message.
 #[test]
 fn note_on() {
-    let mut parser = MidiStreamParser::<256>::new();
+    let mut parser = MidiStreamParser::new();
 
     let bytes = [0x94, 75, 82];
-    let messages = [None, None, Some([0x94, 75, 82].as_ref())];
+    let messages = [None, None, Some(ParserOutput::Message(&[0x94, 75, 82]))];
 
     for (byte, message) in bytes.iter().zip(messages.iter()) {
         let result = parser.parse(*byte).unwrap();
@@ -53,15 +53,15 @@ fn note_on() {
 /// Two note on messages sharing the same status byte.
 #[test]
 fn note_on_running_status() {
-    let mut parser = MidiStreamParser::<256>::new();
+    let mut parser = MidiStreamParser::new();
 
     let bytes = [0x90, 60, 127, 61, 40];
     let messages = [
         None,
         None,
-        Some([0x90, 60, 127].as_ref()),
+        Some(ParserOutput::Message(&[0x90, 60, 127])),
         None,
-        Some([0x90, 61, 40].as_ref()),
+        Some(ParserOutput::Message(&[0x90, 61, 40])),
     ];
 
     for (byte, message) in bytes.iter().zip(messages.iter()) {
@@ -73,10 +73,10 @@ fn note_on_running_status() {
 /// Poly key pressure message.
 #[test]
 fn poly_key_pressure() {
-    let mut parser = MidiStreamParser::<256>::new();
+    let mut parser = MidiStreamParser::new();
 
     let bytes = [0xA1, 52, 80];
-    let messages = [None, None, Some([0xA1, 52, 80].as_ref())];
+    let messages = [None, None, Some(ParserOutput::Message(&[0xA1, 52, 80]))];
 
     for (byte, message) in bytes.iter().zip(messages.iter()) {
         let result = parser.parse(*byte).unwrap();
@@ -87,10 +87,10 @@ fn poly_key_pressure() {
 /// Control change message.
 #[test]
 fn control_change() {
-    let mut parser = MidiStreamParser::<256>::new();
+    let mut parser = MidiStreamParser::new();
 
     let bytes = [0xBC, 21, 40];
-    let messages = [None, None, Some([0xBC, 21, 40].as_ref())];
+    let messages = [None, None, Some(ParserOutput::Message(&[0xBC, 21, 40]))];
 
     for (byte, message) in bytes.iter().zip(messages.iter()) {
         let result = parser.parse(*byte).unwrap();
@@ -101,10 +101,10 @@ fn control_change() {
 /// Program change message.
 #[test]
 fn program_change() {
-    let mut parser = MidiStreamParser::<256>::new();
+    let mut parser = MidiStreamParser::new();
 
     let bytes = [0xC5, 17];
-    let messages = [None, Some([0xC5, 17].as_ref())];
+    let messages = [None, Some(ParserOutput::Message(&[0xC5, 17]))];
 
     for (byte, message) in bytes.iter().zip(messages.iter()) {
         let result = parser.parse(*byte).unwrap();
@@ -115,10 +115,10 @@ fn program_change() {
 /// Channel pressure message.
 #[test]
 fn channel_pressure() {
-    let mut parser = MidiStreamParser::<256>::new();
+    let mut parser = MidiStreamParser::new();
 
     let bytes = [0xD7, 102];
-    let messages = [None, Some([0xD7, 102].as_ref())];
+    let messages = [None, Some(ParserOutput::Message(&[0xD7, 102]))];
 
     for (byte, message) in bytes.iter().zip(messages.iter()) {
         let result = parser.parse(*byte).unwrap();
@@ -129,10 +129,10 @@ fn channel_pressure() {
 /// Pitch bend message.
 #[test]
 fn pitch_bend() {
-    let mut parser = MidiStreamParser::<256>::new();
+    let mut parser = MidiStreamParser::new();
 
     let bytes = [0xE9, 94, 112];
-    let messages = [None, None, Some([0xE9, 94, 112].as_ref())];
+    let messages = [None, None, Some(ParserOutput::Message(&[0xE9, 94, 112]))];
 
     for (byte, message) in bytes.iter().zip(messages.iter()) {
         let result = parser.parse(*byte).unwrap();
@@ -143,10 +143,10 @@ fn pitch_bend() {
 /// MTC quarter frame message.
 #[test]
 fn mtc_quarter_frame() {
-    let mut parser = MidiStreamParser::<256>::new();
+    let mut parser = MidiStreamParser::new();
 
     let bytes = [0xF1, 54];
-    let messages = [None, Some([0xF1, 54].as_ref())];
+    let messages = [None, Some(ParserOutput::Message(&[0xF1, 54]))];
 
     for (byte, message) in bytes.iter().zip(messages.iter()) {
         let result = parser.parse(*byte).unwrap();
@@ -157,10 +157,10 @@ fn mtc_quarter_frame() {
 /// Song position pointer message.
 #[test]
 fn song_position_pointer() {
-    let mut parser = MidiStreamParser::<256>::new();
+    let mut parser = MidiStreamParser::new();
 
     let bytes = [0xF2, 19, 78];
-    let messages = [None, None, Some([0xF2, 19, 78].as_ref())];
+    let messages = [None, None, Some(ParserOutput::Message(&[0xF2, 19, 78]))];
 
     for (byte, message) in bytes.iter().zip(messages.iter()) {
         let result = parser.parse(*byte).unwrap();
@@ -171,10 +171,10 @@ fn song_position_pointer() {
 /// Song select message.
 #[test]
 fn song_select() {
-    let mut parser = MidiStreamParser::<256>::new();
+    let mut parser = MidiStreamParser::new();
 
     let bytes = [0xF3, 12];
-    let messages = [None, Some([0xF3, 12].as_ref())];
+    let messages = [None, Some(ParserOutput::Message(&[0xF3, 12]))];
 
     for (byte, message) in bytes.iter().zip(messages.iter()) {
         let result = parser.parse(*byte).unwrap();
@@ -185,10 +185,10 @@ fn song_select() {
 /// Tune request message.
 #[test]
 fn tune_request() {
-    let mut parser = MidiStreamParser::<256>::new();
+    let mut parser = MidiStreamParser::new();
 
     let bytes = [0xF6];
-    let messages = [Some([0xF6].as_ref())];
+    let messages = [Some(ParserOutput::Message(&[0xF6]))];
 
     for (byte, message) in bytes.iter().zip(messages.iter()) {
         let result = parser.parse(*byte).unwrap();
@@ -199,16 +199,16 @@ fn tune_request() {
 /// Multiple system realtime messages.
 #[test]
 fn system_realtime() {
-    let mut parser = MidiStreamParser::<256>::new();
+    let mut parser = MidiStreamParser::new();
 
     let bytes = [0xF8, 0xFA, 0xFB, 0xFC, 0xFE, 0xFF];
     let messages = [
-        Some([0xF8].as_ref()),
-        Some([0xFA].as_ref()),
-        Some([0xFB].as_ref()),
-        Some([0xFC].as_ref()),
-        Some([0xFE].as_ref()),
-        Some([0xFF].as_ref()),
+        Some(ParserOutput::Message(&[0xF8])),
+        Some(ParserOutput::Message(&[0xFA])),
+        Some(ParserOutput::Message(&[0xFB])),
+        Some(ParserOutput::Message(&[0xFC])),
+        Some(ParserOutput::Message(&[0xFE])),
+        Some(ParserOutput::Message(&[0xFF])),
     ];
 
     for (byte, message) in bytes.iter().zip(messages.iter()) {
@@ -220,16 +220,16 @@ fn system_realtime() {
 /// SysEx message without anything special.
 #[test]
 fn sysex() {
-    let mut parser = MidiStreamParser::<256>::new();
+    let mut parser = MidiStreamParser::new();
 
     let bytes = [0xF0, 0x10, 0x20, 0x7F, 0x30, 0xF7];
     let messages = [
-        None,
-        None,
-        None,
-        None,
-        None,
-        Some([0xF0, 0x10, 0x20, 0x7F, 0x30, 0xF7].as_ref()),
+        Some(ParserOutput::SysexByte(0xF0)),
+        Some(ParserOutput::SysexByte(0x10)),
+        Some(ParserOutput::SysexByte(0x20)),
+        Some(ParserOutput::SysexByte(0x7F)),
+        Some(ParserOutput::SysexByte(0x30)),
+        Some(ParserOutput::SysexByte(0xF7)),
     ];
 
     for (byte, message) in bytes.iter().zip(messages.iter()) {
@@ -241,54 +241,21 @@ fn sysex() {
 /// SysEx message with a clock message in-between.
 #[test]
 fn sysex_with_realtime() {
-    let mut parser = MidiStreamParser::<256>::new();
+    let mut parser = MidiStreamParser::new();
 
     let bytes = [0xF0, 0x10, 0xF8, 0x20, 0x7F, 0x30, 0xF7];
     let messages = [
-        None,
-        None,
-        Some([0xF8].as_ref()),
-        None,
-        None,
-        None,
-        Some([0xF0, 0x10, 0x20, 0x7F, 0x30, 0xF7].as_ref()),
+        Some(ParserOutput::SysexByte(0xF0)),
+        Some(ParserOutput::SysexByte(0x10)),
+        Some(ParserOutput::Message(&[0xF8])),
+        Some(ParserOutput::SysexByte(0x20)),
+        Some(ParserOutput::SysexByte(0x7F)),
+        Some(ParserOutput::SysexByte(0x30)),
+        Some(ParserOutput::SysexByte(0xF7)),
     ];
 
     for (byte, message) in bytes.iter().zip(messages.iter()) {
         let result = parser.parse(*byte).unwrap();
         assert_eq!(result, *message);
-    }
-}
-
-/// SysEx message with more bytes than parser can buffer,
-/// followed by a shorter one that can be processed.
-#[test]
-fn sysex_overflow() {
-    let mut parser = MidiStreamParser::<4>::new();
-
-    let bytes = [0xF0, 0x01, 0x02, 0x03, 0x04, 0xF7, 0xF0, 0x11, 0x12, 0xF7];
-    let messages = [
-        None,
-        None,
-        None,
-        None,
-        None,
-        None,
-        None,
-        None,
-        None,
-        Some([0xF0, 0x11, 0x12, 0xF7].as_ref()),
-    ];
-
-    for (byte, message) in bytes.iter().zip(messages.iter()) {
-        let result = parser.parse(*byte);
-        match result {
-            Ok(result) => {
-                assert_eq!(result, *message);
-            }
-            Err(result) => {
-                assert!(matches!(result, ParserError::SysexOverflow));
-            }
-        }
     }
 }